pyproject.toml
[tool.poetry]
name = "game-data-analyzer"
version = "0.1.0"
description = ""
authors = ["SirAbhi13 <abhinav.sny.2002@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
django-shell-plus = "^1.1.7"
django-extensions = "^3.2.3"
pandas = "^2.2.2"
requests = "^2.32.3"
django-celery-results = "^2.5.1"
sqlalchemy = "^2.0.31"
django-cors-headers = "^4.4.0"
gunicorn = "^22.0.0"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
black = "^24.4.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
